requirements.txt
africastalking==1.2.5
Django==4.2
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
Pillow==9.5.0
requests==2.29.0
mysqlclient==2.1.1
